Removed caching in some pages, fixed build errors, upgraded prisma
