+++ package.json
@@ -37,7 +37,7 @@
     "@mui/lab": "^5.0.0-alpha.146",
     "@mui/material": "^5.14.11",
     "@mui/x-data-grid": "^6.15.0",
-    "@prisma/client": "^5.3.1",
+    "@prisma/client": "^5.7.1",
     "bcrypt": "^5.1.1",
     "google-auth-library": "^9.4.1",
     "googleapis": "^126.0.1",
@@ -64,7 +64,7 @@
     "husky": "^8.0.3",
     "lint-staged": "^14.0.1",
     "prettier": "^3.0.3",
-    "prisma": "^5.3.1",
+    "prisma": "^5.7.1",
     "typescript": "latest"
   }
 }
