+++ src/app/events/[eventId]/page.tsx
@@ -17,3 +17,6 @@ export default async function EventPage({
 
   return <UnderConstruction />;
 }
+
+// Opt out of caching for all data requests in the route segment
+export const dynamic = "force-dynamic";
