+++ src/app/leaderboard/page.tsx
@@ -26,3 +26,6 @@ export default async function HomePage() {
     </Box>
   );
 }
+
+// Opt out of caching for all data requests in the route segment
+export const dynamic = "force-dynamic";
